much stuff changed

+++ sh/scrub.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
 # Log sha1 hashsum for every file in current location
 
-dstmangl=$(echo "$PWD/$1" | tr -d "./ "  )
+dstmangl=$(echo "$1" | tr -d "./ "  )
 export output=/tmp/scrub_"$dstmangl"
 
 echo Output_ $output
-[[ -f "$output" ]] &&
-    rm "$output"
+[[ -f "$output" ]] && rm "$output"
 
 find -L  "$1"  -depth  -type f -not -path '*/\.git/*'  -print 1>"$output"
 
@@ -21,6 +20,7 @@ sort -o "$output" "$output"
 du -h "$output"
 
 echo "Gonna calc sha1 sums.."
+[[ -f "$output"_sums ]] && rm "$output"_sums
 cat "$output" | while read fname
 do
     sum=$(sha1sum "$fname"| cut -f1 -d " ")
@@ -28,5 +28,6 @@ do
 done
 
 mv "$output"_sums "$output"
+echo -n "see result at "
 du -h "$output"
